feat: Introduce web application, including a landing page, shared UI components, and utility hooks.
