+++ apps/web/modules/landing/LandingPage.tsx
@@ -1,20 +1,38 @@
-import SplashCursor from "../../../../packages/ui/src/components/splash-cursor";
-import FeaturesSection from "./components/FeaturesSection";
-import Footer from "./components/Footer";
-import HeroSection from "./components/HeroSection";
+"use client";
+
+import dynamic from "next/dynamic";
+import { useMobileDetect } from "@/hooks/use-mobile-detect";
+
+// Lazy load heavy components for better code splitting and performance
+// SplashCursor is the heaviest (WebGL fluid simulation), only load on medium+ devices
+const SplashCursor = dynamic(
+  () => import("../../../../packages/ui/src/components/splash-cursor"),
+  { 
+    ssr: false, // Disable SSR for client-only WebGL component
+    loading: () => null // No loading state needed
+  }
+);
+
+// Lazy load sections for progressive loading
+const HeroSection = dynamic(() => import("./components/HeroSection"));
+const ProblemSection = dynamic(() => import("./components/ProblemSection"));
+const SolutionSection = dynamic(() => import("./components/SolutionSection"));
+const FeaturesSection = dynamic(() => import("./components/FeaturesSection"));
+const Footer = dynamic(() => import("./components/Footer"));
+
+// Navbar loads immediately (above the fold and critical)
 import Navbar from "./components/Navbar";
-import ProblemSection from "./components/ProblemSection";
-import SolutionSection from "./components/SolutionSection";
 
 export default function LandingPage() {
+  const { isMobile } = useMobileDetect();
+
   return (
     <main className="min-h-screen ">
-      <SplashCursor />
-        <Navbar />
-        <HeroSection />
-        <ProblemSection />
-        
-      
+      {/* Only render SplashCursor on medium+ devices (≥768px) for performance */}
+      {!isMobile && <SplashCursor />}
+      <Navbar />
+      <HeroSection />
+      <ProblemSection />
       <SolutionSection />
       <FeaturesSection />
       <Footer />
