+++ apps/web/next.config.mjs
@@ -1,6 +1,30 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   transpilePackages: ["@workspace/ui"],
-}
+  
+  // Performance optimizations
+  compiler: {
+    // Remove console.log in production
+    removeConsole: process.env.NODE_ENV === "production" ? {
+      exclude: ["error", "warn"],
+    } : false,
+  },
 
-export default nextConfig
+  // Optimize images
+  images: {
+    formats: ["image/webp", "image/avif"],
+  },
+
+  // Enable experimental features for better performance
+  experimental: {
+    optimizePackageImports: [
+      "@workspace/ui",
+      "motion",
+    ],
+  },
+
+  // Production optimizations
+  reactStrictMode: true,
+};
+
+export default nextConfig;
