+++ packages/ui/src/components/topography-pattern.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+
+/**
+ * TopographyPattern - SVG-based topographical background pattern
+ * Creates a technical, map-like aesthetic
+ */
+export const TopographyPattern = React.memo(() => {
+  return (
+    <div className="absolute inset-0 -z-10 h-full w-full">
+      {/* SVG Pattern Definition */}
+      <div
+        className="absolute inset-0 h-full w-full [mask-image:radial-gradient(ellipse_at_center,white_40%,transparent_80%)] opacity-[0.1] dark:opacity-[0.2]"
+        style={{
+          backgroundImage: `url("data:image/svg+xml,%3Csvg width='60' height='60' viewBox='0 0 60 60' xmlns='http://www.w3.org/2000/svg'%3E%3Cg fill='none' fill-rule='evenodd'%3E%3Cg fill='%239C92AC' fill-opacity='1'%3E%3Cpath d='M36 34v-4h-2v4h-4v2h4v4h2v-4h4v-2h-4zm0-30V0h-2v4h-4v2h4v4h2V6h4V4h-4zM6 34v-4H4v4H0v2h4v4h2v-4h4v-2H6zM6 4V0H4v4H0v2h4v4h2V6h4V4H6z'/%3E%3C/g%3E%3C/g%3E%3C/svg%3E")`,
+        }}
+      />
+    </div>
+  );
+});
+
+TopographyPattern.displayName = "TopographyPattern";
